+++ HISTORY.rst
@@ -13,6 +13,7 @@ History
 - Apply `form-check`, `form-check-label` and `form-check-input` classes to appropriate tags for `RadioSelect` and `CheckboxSelectMultiple` widgets (#141)
 - Errors on file inputs are shown (#146)
 - Only display non-field errors in form alert block (#161)
+- Reinstate `bootstrap4_title` block to `bootstrap4.html` (#156)
 
 0.0.8 (2019-03-12)
 ++++++++++++++++++